factor out reference-fasta args
